pyproject.toml
[project]
name = "athena"
version = "1.0.0"
description = "Athena — Your personal AI professor that reads your PDFs and teaches you"
authors = [{name = "Harsh Rohilla", email = "harshrohilla9876@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "fastmcp>=2.13.3",
    "langchain>=1.1.2",
    "langchain-text-splitters>=1.0.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pinecone>=8.0.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
